+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Show filter syntax guide on empty page
 - Add syntax highlighting for filters like `album:"mirai sekai"`
 - Add `artist:` filtering for titles without brackets, like `Supercell feat. Veela`
+- Add `artist:` filtering for title prefixes `w/`, `prod`, `prod by` and suffixes `cover`, `rework`, `refix`, `vip`, `mashup`
 - Remove the active filter highlight
 - Fix undo/redo in the filter textbox only changing one character at a time
 - Fix "electron.app.hide is not a function" error when closing app window on Windows/Linux